@@ -42,25 +42,25 @@
   },
   "homepage": "https://github.com/ghostery/ghostery-extension#readme",
   "dependencies": {
-    "@cliqz/adblocker-circumvention": "^1.0.2",
+    "@cliqz/adblocker-circumvention": "^1.2.0",
     "@cliqz/url-parser": "^1.0.3",
     "base64-js": "^1.3.1",
     "browser-core": "https://github.com/cliqz-oss/browser-core/releases/download/v7.39.5/browser-core-7.39.5.tgz",
     "classnames": "^2.2.5",
-    "d3": "^5.11.0",
+    "d3": "^5.12.0",
     "foundation-sites": "^6.4.4-rc1",
-    "history": "^4.9.0",
+    "history": "^4.10.1",
     "json-api-normalizer": "^0.4.16",
     "moment": "^2.19.1",
     "prop-types": "^15.6.2",
     "query-string": "^6.8.3",
-    "react": "^16.9.0",
-    "react-dom": "^16.9.0",
+    "react": "^16.10.1",
+    "react-dom": "^16.10.1",
     "react-markdown": "^4.2.2",
     "react-redux": "^7.1.1",
-    "react-router-dom": "^5.0.1",
+    "react-router-dom": "^5.1.2",
     "react-shadow-dom-retarget-events": "^1.0.10",
-    "react-svg": "^10.0.20",
+    "react-svg": "^10.0.22",
     "redux": "^4.0.4",
     "redux-object": "^0.5.10",
     "redux-thunk": "^2.2.0",
@@ -71,25 +71,25 @@
     "whatwg-fetch": "^3.0.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.5.5",
+    "@babel/core": "^7.6.2",
     "@babel/plugin-proposal-class-properties": "^7.5.5",
-    "@babel/plugin-proposal-object-rest-spread": "^7.5.5",
-    "@babel/plugin-transform-modules-commonjs": "^7.5.0",
+    "@babel/plugin-proposal-object-rest-spread": "^7.6.2",
+    "@babel/plugin-transform-modules-commonjs": "^7.6.0",
     "@babel/preset-react": "^7.0.0",
-    "@babel/register": "^7.5.5",
+    "@babel/register": "^7.6.2",
     "babel-eslint": "^10.0.3",
     "babel-loader": "^8.0.6",
     "clean-webpack-plugin": "^3.0.0",
-    "cross-env": "^5.2.1",
+    "cross-env": "^6.0.3",
     "css-loader": "^3.2.0",
     "enzyme": "^3.10.0",
     "enzyme-adapter-react-16": "^1.14.0",
-    "eslint": "^6.3.0",
+    "eslint": "^6.5.1",
     "eslint-config-airbnb": "^18.0.1",
-    "eslint-loader": "^3.0.0",
+    "eslint-loader": "^3.0.2",
     "eslint-plugin-import": "^2.18.2",
     "eslint-plugin-jsx-a11y": "^6.2.3",
-    "eslint-plugin-react": "^7.14.3",
+    "eslint-plugin-react": "^7.15.1",
     "fs-extra": "^8.1.0",
     "jest": "^24.9.0",
     "jsdoc": "^3.6.3",
@@ -99,19 +99,19 @@
     "node-sass": "^4.12.0",
     "oboe": "^2.1.3",
     "path": "^0.12.7",
-    "react-router": "^5.0.1",
-    "react-test-renderer": "^16.9.0",
+    "react-router": "^5.1.2",
+    "react-test-renderer": "^16.10.1",
     "redux-mock-store": "^1.5.1",
     "sass-loader": "^8.0.0",
     "seamless-immutable": "^7.1.3",
-    "sinon": "^7.4.2",
+    "sinon": "^7.5.0",
     "sinon-chrome": "^3.0.1",
     "svg-url-loader": "^3.0.2",
     "underscore-template-loader": "^1.0.0",
     "url-loader": "^2.1.0",
     "vendor-copy": "^2.0.0",
-    "webpack": "^4.39.3",
-    "webpack-cli": "^3.3.8",
+    "webpack": "^4.41.0",
+    "webpack-cli": "^3.3.9",
     "webpack-shell-plugin": "^0.5.0"
   }
 }
